Use Dataflow template for reading Spanner change streams
