--- pipelines/etl_integration_java/scripts/03_run_changestream_dataflow.sh
@@ -1,20 +1,0 @@
-./gradlew run -Pargs="
-  --pipeline=SPANNER_CHANGE_STREAM \
-  --streaming \
-  --enableStreamingEngine \
-  --autoscalingAlgorithm=THROUGHPUT_BASED \
-  --runner=DataflowRunner \
-  --experiments=use_runner_v2 \
-  --project=$PROJECT \
-  --tempLocation=$TEMP_LOCATION \
-  --region=$REGION \
-  --serviceAccount=$SERVICE_ACCOUNT \
-  --subnetwork=$NETWORK \
-  --maxNumWorkers=$MAX_DATAFLOW_WORKERS \
-  --experiments=enable_data_sampling;use_network_tags=ssh;dataflow \
-  --usePublicIps=false \
-  --spannerInstance=$SPANNER_INSTANCE \
-  --spannerDatabase=$SPANNER_DATABASE \
-  --spannerChangeStream=$SPANNER_CHANGE_STREAM \
-  --bigQueryDataset=$BIGQUERY_DATASET \
-  --bigQueryTable=$SPANNER_TABLE"
