+++ pipelines/etl_integration_java/src/main/java/com/google/cloud/dataflow/solutions/data/TaxiObjects.java
@@ -16,10 +16,7 @@
 
 package com.google.cloud.dataflow.solutions.data;
 
-import com.google.api.services.bigquery.model.TableRow;
 import com.google.auto.value.AutoValue;
-import java.util.List;
-import javax.annotation.Nullable;
 import org.apache.beam.sdk.schemas.AutoValueSchema;
 import org.apache.beam.sdk.schemas.annotations.DefaultSchema;
 import org.apache.beam.sdk.schemas.annotations.SchemaFieldName;
@@ -84,67 +81,6 @@ public class TaxiObjects {
             public abstract Builder setPassengerCount(Integer value);
 
             public abstract TaxiEvent build();
-        }
-
-        public static List<String> primaryKeys() {
-            return List.of("ride_id", "point_idx");
-        }
-    }
-
-    @DefaultSchema(AutoValueSchema.class)
-    @AutoValue
-    public abstract static class CDCValue {
-        @SchemaFieldName("mod_type")
-        public abstract String getModType();
-
-        @SchemaFieldName("event")
-        public abstract @Nullable TaxiEvent getEvent();
-
-        @SchemaFieldName("sequence_number")
-        public abstract Long getSequenceNumber();
-
-        public static Builder builder() {
-            return new AutoValue_TaxiObjects_CDCValue.Builder();
-        }
-
-        @AutoValue.Builder
-        public abstract static class Builder {
-            public abstract Builder setModType(String value);
-
-            public abstract Builder setEvent(TaxiEvent value);
-
-            public abstract Builder setSequenceNumber(Long value);
-
-            public abstract CDCValue build();
-        }
-    }
-
-    @DefaultSchema(AutoValueSchema.class)
-    @AutoValue
-    public abstract static class CDCValueForBQ {
-        @SchemaFieldName("event")
-        public abstract TableRow getTableRow();
-
-        @SchemaFieldName("sequence_number")
-        public abstract Long getSequenceNumber();
-
-        @SchemaFieldName("mod_type")
-        public abstract String getModType();
-
-        public static Builder builder() {
-            return new AutoValue_TaxiObjects_CDCValueForBQ.Builder();
-        }
-
-        @AutoValue.Builder
-        public abstract static class Builder {
-
-            public abstract Builder setTableRow(TableRow value);
-
-            public abstract Builder setSequenceNumber(Long value);
-
-            public abstract Builder setModType(String value);
-
-            public abstract CDCValueForBQ build();
         }
     }
