--- pipelines/etl_integration_java/src/main/java/com/google/cloud/dataflow/solutions/options/RunOptions.java
@@ -1,34 +1,0 @@
-/*
-*  Copyright 2024 Google LLC
-*
-*  Licensed under the Apache License, Version 2.0 (the "License");
-*  you may not use this file except in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*      https://www.apache.org/licenses/LICENSE-2.0
-*
-*  Unless required by applicable law or agreed to in writing, software
-*  distributed under the License is distributed on an "AS IS" BASIS,
-*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-*  See the License for the specific language governing permissions and
-*  limitations under the License.
-*/
-
-package com.google.cloud.dataflow.solutions.options;
-
-import org.apache.beam.sdk.options.Description;
-import org.apache.beam.sdk.options.PipelineOptions;
-import org.apache.beam.sdk.options.Validation;
-
-public interface RunOptions extends PipelineOptions {
-    enum PipelineToRun {
-        PUBSUB_TO_SPANNER,
-        SPANNER_CHANGE_STREAM
-    }
-
-    @Description("Name of the pipeline to be run")
-    @Validation.Required()
-    void setPipeline(PipelineToRun p);
-
-    PipelineToRun getPipeline();
-}
